% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_estim.R
\name{is_estim}
\alias{is_estim}
\title{Perform importance sampling}
\usage{
is_estim(N, se, omega_inv_sq, ce, out, ncores = 1)
}
\arguments{
\item{N}{An integer, number of simulations to perform}

\item{se}{A vector of effect size standard errors}

\item{omega_inv_sq}{A matrix of square root inverse of genetic correlation}

\item{ce}{A matrix of the genetic correlation}

\item{out}{A string giving the output file}

\item{ncores}{An integer of the number of cores}
}
\value{
none
}
\description{
Perform importance sampling
}
