% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleio.R
\name{pleio}
\alias{pleio}
\title{Perform RPleio}
\usage{
pleio(sum_file, g_file, e_file, nis, ncores, isf = "", out)
}
\arguments{
\item{sum_file}{A string giving the summary statistic file}

\item{g_file}{A string giving the genetic correlation matrix file}

\item{e_file}{A string giving the environmental correlation matrix file}

\item{nis}{A integer giving the number of samples to be used for importance sampling}

\item{ncores}{An integer giing the number of cores to use for parallelization}

\item{isf}{A string giving the importance sampling file. Optional}

\item{out}{A string giving the output prefix}
}
\value{
none
}
\description{
Perform RPleio
}
